remove drakeGeometryUtil
