+++ drake/systems/plants/constraint/RigidBodyConstraint.cpp
@@ -6,9 +6,9 @@
 #include "drake/math/autodiff.h"
 #include "drake/math/autodiff_gradient.h"
 #include "drake/math/gradient.h"
+#include "drake/math/geometry.h"
 #include "drake/math/quaternion.h"
 #include "drake/systems/plants/RigidBodyTree.h"
-#include "drake/util/drakeGeometryUtil.h"
 
 using namespace Eigen;
