+++ drake/systems/robotInterfaces/QPLocomotionPlan.cpp
@@ -12,11 +12,11 @@
 #include "drake/math/autodiff.h"
 #include "drake/math/autodiff_gradient.h"
 #include "drake/math/expmap.h"
+#include "drake/math/geometry.h"
 #include "drake/math/gradient.h"
 #include "drake/math/quaternion.h"
 #include "drake/math/rotation_matrix.h"
 #include "drake/solvers/qpSpline/splineGeneration.h"
-#include "drake/util/drakeGeometryUtil.h"
 #include "drake/util/drakeUtil.h"
 #include "drake/util/lcmUtil.h"
 #include "drake/util/convexHull.h"
